Updated when next steps is accessible
