+++ src/backend/src/modules/pia-intake/dto/create-pia-intake.dto.ts
@@ -140,24 +140,6 @@ export class CreatePiaIntakeDto {
   })
   submittedAt: Date;
 
-  @IsBoolean()
-  @IsOptional()
-  @ApiProperty({
-    type: Boolean,
-    required: false,
-    example: piaIntakeEntityMock.isNextStepsSeenForDelegatedFlow,
-  })
-  isNextStepsSeenForDelegatedFlow: boolean;
-
-  @IsBoolean()
-  @IsOptional()
-  @ApiProperty({
-    type: Boolean,
-    required: false,
-    example: piaIntakeEntityMock.isNextStepsSeenForNonDelegatedFlow,
-  })
-  isNextStepsSeenForNonDelegatedFlow: boolean;
-
   @IsObject()
   @IsOptional()
   @IsNotEmptyObject()
