+++ src/backend/src/modules/pia-intake/metadata/pia-status.metadata.ts
@@ -107,8 +107,6 @@ export const piaStatusMetadata: Partial<
       overrides: {
         pia: {
           review: true,
-          isNextStepsSeenForDelegatedFlow: true,
-          isNextStepsSeenForNonDelegatedFlow: true,
         },
         comment: {
           add: true,
