+++ src/backend/test/util/mocks/data/pia-intake.mock.ts
@@ -33,8 +33,6 @@ const piaIntakeDataMock = {
   saveId: 1,
   submittedAt: new Date(),
   isActive: true,
-  isNextStepsSeenForDelegatedFlow: false,
-  isNextStepsSeenForNonDelegatedFlow: false,
   collectionUseAndDisclosure: {
     steps: [
       {
