+++ src/frontend/src/components/public/PIAFormTabs/nextSteps/helper/interfaces.ts
@@ -17,7 +17,3 @@ export interface IModalObject {
     statusChange: PiaStatuses;
   };
 }
-
-export interface PIFlow {
-  navigateFn: (url: string) => void;
-}
