+++ src/frontend/src/components/public/PIANavButton/interface.ts
@@ -2,6 +2,4 @@ import { INavbarItem } from '../../common/Navbar/interfaces';
 
 export interface INavButton {
   pages: INavbarItem[];
-  isDelegate: boolean;
-  isIntakeSubmitted: boolean;
 }
