+++ src/frontend/src/routes/router.tsx
@@ -3,7 +3,7 @@ import NotFound from '../components/common/NotFound';
 import Unauthorized from '../components/common/Unauthorized';
 import Forbidden from '../components/common/Forbidden';
 import { PIAFormIntake } from '../components/public/PIAFormTabs/intake';
-import { PIANextSteps } from '../components/public/PIAFormTabs/Next_Steps';
+import { PIANextSteps } from '../components/public/PIAFormTabs/nextSteps';
 import LandingPage from '../pages/LandingPage/LandingPage';
 import PIAFormPage from '../pages/PIAForm';
 import PIAIntakeResultsPage from '../pages/PIAIntakeResultsPage/PIAIntakeResultsPage';
